+++ R/checks-base.R
@@ -58,10 +58,49 @@ is_variable_check_func <- function(func) {
   identical(arg_ast$head, "$")
 }
 
+check_errors_attr <- "check.errors"
+
+has_check_errors <- function(x) {
+  has_attr(x, check_errors_attr)
+}
+
+get_check_errors <- function(x) {
+  get_attr(x, check_errors_attr)
+}
+
+set_check_errors <- function(x, reasons) {
+  set_attr(x, check_errors_attr, reasons)
+}
+
+#' Helper function to embed reasons
+#' @noRd
+fail_check <- function(reasons = NULL) {
+  set_check_errors(FALSE, reasons)
+}
+
+#' Helper function to embed reasons
+#' @noRd
+warn_check <- function(reasons = NULL) {
+  set_check_errors(TRUE, reasons)
+}
+
 #' Evaluate all checks on a blueprint
 #'
 #' Runs all checks -- dataset and variable -- on a blueprint
 #' to determine if a built dataset passes all restrictions.
+#'
+#' @details # Check functions
+#'
+#' Check functions are simple functions that take in either
+#' a data.frame or variable at the minimum, plus some extra
+#' arguments if need, and returns a logical value: `TRUE` or `FALSE.`
+#' In blueprintr, the entire check passes or fails unlike other
+#' testing frameworks like {pointblank}. If you'd like to embed
+#' extra context for your test result, modify the "check.errors"
+#' attribute of the returned logical value with a character vector
+#' which will be rendered into a bulleted list. Note: if you embed
+#' reasons for a `TRUE`, the check will produce a warning in the drake
+#' or targets pipeline.
 #'
 #' @param ... All quoted check calls
 #' @param .env The environment in which the calls are evaluated
@@ -70,7 +109,21 @@ is_variable_check_func <- function(func) {
 eval_checks <- function(..., .env = parent.frame()) {
   checks_dt <- checks(...)
 
-  checks_dt$.pass <- vlapply(checks_dt$check_func, function(f) eval(f, envir = .env))
+  checks_results <- lapply(checks_dt$check_func, function(f) eval(f, envir = .env))
+  checks_dt$.pass <- vlapply(checks_results, as.logical)
+  checks_dt$.fail_meta <- vcapply(checks_results, \(x) {
+    if (has_check_errors(x)) {
+      errs <- get_check_errors(x)
+      errs <- vcapply(errs, \(x) paste0("  * ", x))
+      paste0(":\n", paste0(errs, collapse = "\n"))
+    } else {
+      ""
+    }
+  })
+
+  if (any(checks_dt$.pass == TRUE & checks_dt$.fail_meta != "")) {
+    checks_warn(checks_dt)
+  }
 
   if (any(checks_dt$.pass == FALSE)) {
     checks_error(checks_dt)
@@ -80,18 +133,62 @@ eval_checks <- function(..., .env = parent.frame()) {
 }
 
 checks_error <- function(checks) {
-  false_funcs <-
-    checks %>%
-    dplyr::filter(.data$.pass == FALSE) %>%
-    dplyr::pull(.data$check_func) %>%
-    vcapply(safe_deparse, collapse = " ", trim = TRUE)
+  checks <- checks %>%
+    dplyr::mutate(
+      .call = vcapply(.data$check_func, safe_deparse, collapse = " ", trim = TRUE),
+    ) %>%
+    dplyr::mutate(
+      .message = dplyr::if_else(
+        .data$.pass == FALSE,
+        paste0(
+          "`", .data$.call, "` is not TRUE", .data$.fail_meta
+        ),
+        NA_character_
+      ),
+      # Include the warning here for user experience
+      .message = dplyr::if_else(
+        .data$.pass == TRUE & .data$.fail_meta != "",
+        paste0(
+          "`", .data$.call, "` has some potential issues", .data$.fail_meta
+        ),
+        .data$.message
+      )
+    )
 
-  err_msgs <- glue("`{false_funcs}` is not TRUE")
+  err_msgs <- checks %>%
+    dplyr::filter(!is.na(.data$.message)) %>%
+    dplyr::pull(.data$.message)
 
   rlang::abort(
     glue_collapse(err_msgs, "\n"),
-    checks = checks,
-    .subclass = "checks_error"
+    class = "checks_error",
+    checks = checks
+  )
+}
+
+checks_warn <- function(checks) {
+  checks <- checks %>%
+    dplyr::mutate(
+      .call = vcapply(.data$check_func, safe_deparse, collapse = " ", trim = TRUE),
+    ) %>%
+    dplyr::mutate(
+      .message = dplyr::if_else(
+        .data$.pass == TRUE & .data$.fail_meta != "",
+        paste0(
+          "`", .data$.call, "` has some potential issues", .data$.fail_meta
+        ),
+        NA_character_
+      )
+    )
+
+  warn_msgs <- checks %>%
+    dplyr::filter(!is.na(.data$.message)) %>%
+    dplyr::pull(.data$.message)
+
+  rlang::warn(
+    message = glue_collapse(warn_msgs, "\n"),
+    class = "checks_warn",
+    checks = checks
   )
 }